+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     - Added Levenshtein distance kernel
     - Move K Best Selector to main repository
     - Added custom exceptions from the main repo
+    - Moved Flysystem Persister over to main repo
 
 - 0.2.1-beta
     - Implemented K Best feature selector